機能の整理を開始。他の機能の実装が一通り片付いたため
